move run_update_check to utils
