+++ circleguard/utils.py
@@ -1,9 +1,18 @@
 from pathlib import Path
 import sys
 import os
+from datetime import datetime, timedelta
 
+from packaging import version
+import requests
+
+# placeholder imports to have all imports at the top of the file. Imported for
+# real farther below
+#from settings import get_setting
+#from version import __version__
+
+# placed above local imports to avoid circular import errors
 ROOT_PATH = Path(__file__).parent.absolute()
-
 def resource_path(str_path):
     """
     Returns a Path representing where to look for resource files for the program,
@@ -16,6 +25,34 @@ def resource_path(str_path):
         return Path(sys._MEIPASS) / Path(str_path)  # pylint: disable=no-member
     return ROOT_PATH / Path(str_path)
 
+
+from settings import get_setting
+from version import __version__
+
+
+def run_update_check():
+    last_check = datetime.strptime(get_setting("last_update_check"), get_setting("timestamp_format"))
+    next_check = last_check + timedelta(hours=1)
+    if not next_check < datetime.now():
+        return get_idle_setting_str()
+    try:
+        # check for new version
+        git_request = requests.get("https://api.github.com/repos/circleguard/circleguard/releases/latest").json()
+        git_version = version.parse(git_request["name"])
+        update_default("update_version", git_version)
+        update_default("last_update_check", datetime.now().strftime(get_setting("timestamp_format")))
+    except:
+        # user is propably offline
+        pass
+    return get_idle_setting_str()
+
+
+def get_idle_setting_str():
+    current_version = version.parse(__version__)
+    if current_version <= version.parse(get_setting("update_version")):
+        return "<a href=\'https://circleguard.dev/download'>Update available!</a>"
+    else:
+        return "Idle"
 
 class Run():
     """
